+++ utilities/source_utilities.py
@@ -274,10 +274,17 @@ def approve_event(id):
 def disapprove_event(id):
     print("{} is going to be disapproved".format(id))
     result = find_one_and_update(current_app.config['EVENT_COLLECTION'], condition={"_id": ObjectId(id)}, update={
-        "$set": {"eventStatus":  "disapproved"}
+        "$set": {"eventStatus":  "approved"}
     })
     if not result:
         print("Disapprove event {} fails in disapprove_event".format(id))
+    else:
+        if result.get("platformEventId"):
+            objectId_list_to_delete = list()
+            objectId_list_to_delete.append(ObjectId(id))
+            delete_events_in_building_block(objectId_list_to_delete)
+            find_one_and_update(current_app.config['EVENT_COLLECTION'], condition={"_id": ObjectId(id)}, update={
+                "$set": {"submitType": "post", "platformEventId": None}})
 
 
 def get_event(objectId):